Refactor complex import to reduce internal import
