+++ app_clustering.py
@@ -21,7 +21,7 @@ clustering_bp = Blueprint('clustering_bp', __name__)
 
 def clustering_task_failure_handler(job, connection, type, value, tb):
     """A failure handler for the main clustering task, executed by the worker."""
-    from app import app
+    from flask_app import app
     from app_helper import save_task_status, TASK_STATUS_FAILURE
     with app.app_context():
         task_id = getattr(job, 'id', None) or getattr(job, 'get_id', lambda: None)()
