--- tasks/mediaserver.py
@@ -1,413 +1,0 @@
-# tasks/mediaserver.py
-
-import logging
-import os
-import config  # Import the config module to access server type and settings
-
-# Import the specific implementations
-from tasks.mediaserver_jellyfin import (
-    resolve_user as jellyfin_resolve_user,
-    get_all_playlists as jellyfin_get_all_playlists,
-    get_lyrics as jellyfin_get_lyrics,
-    delete_playlist as jellyfin_delete_playlist,
-    get_recent_albums as jellyfin_get_recent_albums,
-    get_tracks_from_album as jellyfin_get_tracks_from_album,
-    search_albums as jellyfin_search_albums,
-    test_connection as jellyfin_test_connection,
-    download_track as jellyfin_download_track,
-    get_all_songs as jellyfin_get_all_songs,
-    get_playlist_by_name as jellyfin_get_playlist_by_name,
-    create_playlist as jellyfin_create_playlist,
-    create_instant_playlist as jellyfin_create_instant_playlist,
-    create_or_replace_playlist as jellyfin_create_or_replace_playlist,
-    get_top_played_songs as jellyfin_get_top_played_songs,
-    get_last_played_time as jellyfin_get_last_played_time,
-    list_libraries as _jellyfin_list_libraries,
-)
-from tasks.mediaserver_navidrome import (
-    get_all_playlists as navidrome_get_all_playlists,
-    get_lyrics as navidrome_get_lyrics,
-    delete_playlist as navidrome_delete_playlist,
-    get_recent_albums as navidrome_get_recent_albums,
-    get_tracks_from_album as navidrome_get_tracks_from_album,
-    search_albums as navidrome_search_albums,
-    test_connection as navidrome_test_connection,
-    download_track as navidrome_download_track,
-    get_all_songs as navidrome_get_all_songs,
-    get_playlist_by_name as navidrome_get_playlist_by_name,
-    create_playlist as navidrome_create_playlist,
-    create_instant_playlist as navidrome_create_instant_playlist,
-    create_or_replace_playlist as navidrome_create_or_replace_playlist,
-    get_top_played_songs as navidrome_get_top_played_songs,
-    get_last_played_time as navidrome_get_last_played_time,
-    list_libraries as _navidrome_list_libraries,
-)
-from tasks.mediaserver_lyrion import (
-    get_all_playlists as lyrion_get_all_playlists,
-    get_lyrics as lyrion_get_lyrics,
-    delete_playlist as lyrion_delete_playlist,
-    get_recent_albums as lyrion_get_recent_albums,
-    get_tracks_from_album as lyrion_get_tracks_from_album,
-    search_albums as lyrion_search_albums,
-    test_connection as lyrion_test_connection,
-    download_track as lyrion_download_track,
-    get_all_songs as lyrion_get_all_songs,
-    get_playlist_by_name as lyrion_get_playlist_by_name,
-    create_playlist as lyrion_create_playlist,
-    create_instant_playlist as lyrion_create_instant_playlist,
-    create_or_replace_playlist as lyrion_create_or_replace_playlist,
-    get_top_played_songs as lyrion_get_top_played_songs,
-    get_last_played_time as lyrion_get_last_played_time,
-    list_libraries as _lyrion_list_libraries,
-)
-from tasks.mediaserver_mpd import (
-    get_all_playlists as mpd_get_all_playlists,
-    delete_playlist as mpd_delete_playlist,
-    get_recent_albums as mpd_get_recent_albums,
-    get_tracks_from_album as mpd_get_tracks_from_album,
-    download_track as mpd_download_track,
-    get_all_songs as mpd_get_all_songs,
-    get_playlist_by_name as mpd_get_playlist_by_name,
-    create_playlist as mpd_create_playlist,
-    create_instant_playlist as mpd_create_instant_playlist,
-    get_top_played_songs as mpd_get_top_played_songs,
-    get_last_played_time as mpd_get_last_played_time,
-)
-from tasks.mediaserver_emby import (
-    resolve_user as emby_resolve_user,
-    get_all_playlists as emby_get_all_playlists,
-    get_lyrics as emby_get_lyrics,
-    delete_playlist as emby_delete_playlist,
-    get_recent_albums as emby_get_recent_albums,
-    get_tracks_from_album as emby_get_tracks_from_album,
-    search_albums as emby_search_albums,
-    test_connection as emby_test_connection,
-    download_track as emby_download_track,
-    get_all_songs as emby_get_all_songs,
-    get_playlist_by_name as emby_get_playlist_by_name,
-    create_playlist as emby_create_playlist,
-    create_instant_playlist as emby_create_instant_playlist,
-    create_or_replace_playlist as emby_create_or_replace_playlist,
-    get_top_played_songs as emby_get_top_played_songs,
-    get_last_played_time as emby_get_last_played_time,
-    list_libraries as _emby_list_libraries,
-)
-
-logger = logging.getLogger(__name__)
-
-
-# ##############################################################################
-# PUBLIC API (Dispatcher functions)
-# ##############################################################################
-
-def resolve_emby_jellyfin_user(identifier, token):
-    """Public dispatcher for resolving a Jellyfin or Emby user identifier."""
-    # This is specific to Jellyfin, so we call it directly.
-    if config.MEDIASERVER_TYPE == 'jellyfin': return jellyfin_resolve_user(identifier, token)
-    if config.MEDIASERVER_TYPE == 'emby': return emby_resolve_user(identifier, token)
-    return []
-
-def _delete_matching_playlists(playlists_to_check, delete_function, suffix):
-    """Deletes every playlist whose name ends with the suffix; keeps going if one deletion fails."""
-    deleted_count = 0
-    for p in playlists_to_check:
-        # Navidrome uses 'id', others use 'Id'. Check for both.
-        playlist_id = p.get('Id') or p.get('id')
-        try:
-            if p.get('Name', '').endswith(suffix) and delete_function(playlist_id):
-                deleted_count += 1
-        except Exception:
-            logger.exception(f"Failed to delete playlist {playlist_id}; continuing with the remaining playlists.")
-    return deleted_count
-
-def delete_playlists_by_suffix(suffix):
-    """Deletes all playlists whose name ends with the given suffix using admin credentials."""
-    logger.info(f"Starting deletion of all '{suffix}' playlists.")
-    deleted_count = 0
-
-    playlists_to_check = []
-    delete_function = None
-
-    if config.MEDIASERVER_TYPE == 'jellyfin':
-        playlists_to_check = jellyfin_get_all_playlists()
-        delete_function = jellyfin_delete_playlist
-    elif config.MEDIASERVER_TYPE == 'navidrome':
-        playlists_to_check = navidrome_get_all_playlists()
-        delete_function = navidrome_delete_playlist
-    elif config.MEDIASERVER_TYPE == 'lyrion':
-        playlists_to_check = lyrion_get_all_playlists()
-        delete_function = lyrion_delete_playlist
-    elif config.MEDIASERVER_TYPE == 'mpd':
-        playlists_to_check = mpd_get_all_playlists()
-        delete_function = mpd_delete_playlist
-    elif config.MEDIASERVER_TYPE == 'emby':
-        playlists_to_check = emby_get_all_playlists()
-        delete_function = emby_delete_playlist
-
-    if delete_function:
-        deleted_count = _delete_matching_playlists(playlists_to_check, delete_function, suffix)
-
-    logger.info(f"Finished deletion. Deleted {deleted_count} playlists.")
-
-def delete_automatic_playlists():
-    """Deletes all playlists ending with '_automatic' using admin credentials."""
-    delete_playlists_by_suffix('_automatic')
-
-def get_recent_albums(limit):
-    """Fetches recently added albums using admin credentials."""
-    if config.MEDIASERVER_TYPE == 'jellyfin': return jellyfin_get_recent_albums(limit)
-    if config.MEDIASERVER_TYPE == 'navidrome': return navidrome_get_recent_albums(limit)
-    if config.MEDIASERVER_TYPE == 'lyrion': return lyrion_get_recent_albums(limit)
-    if config.MEDIASERVER_TYPE == 'mpd': return mpd_get_recent_albums(limit)
-    if config.MEDIASERVER_TYPE == 'emby': return emby_get_recent_albums(limit)
-    return []
-
-def get_tracks_from_album(album_id, user_creds=None, provider_type=None):
-    """Fetches tracks for an album, optionally using explicit creds."""
-    provider_type = provider_type or config.MEDIASERVER_TYPE
-    if provider_type == 'jellyfin': return jellyfin_get_tracks_from_album(album_id, user_creds=user_creds)
-    if provider_type == 'navidrome': return navidrome_get_tracks_from_album(album_id, user_creds=user_creds)
-    if provider_type == 'lyrion': return lyrion_get_tracks_from_album(album_id, user_creds=user_creds)
-    if provider_type == 'mpd': return mpd_get_tracks_from_album(album_id)
-    if provider_type == 'emby': return emby_get_tracks_from_album(album_id, user_creds=user_creds)
-    return []
-
-def download_track(temp_dir, item):
-    """Downloads a track using admin credentials. Detects format from file if .tmp extension is used."""
-    downloaded_path = None
-    
-    if config.MEDIASERVER_TYPE == 'jellyfin': downloaded_path = jellyfin_download_track(temp_dir, item)
-    elif config.MEDIASERVER_TYPE == 'navidrome': downloaded_path = navidrome_download_track(temp_dir, item)
-    elif config.MEDIASERVER_TYPE == 'lyrion': downloaded_path = lyrion_download_track(temp_dir, item)
-    elif config.MEDIASERVER_TYPE == 'mpd': downloaded_path = mpd_download_track(temp_dir, item)
-    elif config.MEDIASERVER_TYPE == 'emby': downloaded_path = emby_download_track(temp_dir, item)
-    
-    # If download failed or returned None, return as is
-    if not downloaded_path:
-        return None
-    
-    # If file has .tmp extension, try to detect real format from file content
-    if downloaded_path.endswith('.tmp'):
-        try:
-            # Check if file exists before trying to detect format
-            if not os.path.exists(downloaded_path):
-                logger.warning(f"Downloaded file does not exist: {downloaded_path}")
-                return downloaded_path
-                
-            detected_ext = _detect_audio_format(downloaded_path)
-            if detected_ext and detected_ext != '.tmp':
-                new_path = downloaded_path.replace('.tmp', detected_ext)
-                # Check if target file already exists (avoid overwriting)
-                if os.path.exists(new_path):
-                    logger.warning(f"Target file already exists, keeping .tmp: {new_path}")
-                    return downloaded_path
-                os.rename(downloaded_path, new_path)
-                logger.info(f"Detected format and renamed: {os.path.basename(downloaded_path)} -> {os.path.basename(new_path)}")
-                return new_path
-        except Exception as e:
-            logger.debug(f"Format detection failed for {os.path.basename(downloaded_path)}, keeping .tmp: {e}")
-    
-    return downloaded_path
-
-
-def _detect_audio_format(filepath):
-    """Detects audio format from file magic numbers. Returns extension like '.mp3' or '.flac'."""
-    try:
-        with open(filepath, 'rb') as f:
-            header = f.read(12)
-            
-            # Check magic numbers for common audio formats
-            if len(header) < 4:
-                return '.tmp'
-            
-            # FLAC: fLaC
-            if header[:4] == b'fLaC':
-                return '.flac'
-            
-            # MP3: ID3 tag or MP3 sync bits
-            if header[:3] == b'ID3' or (len(header) >= 2 and header[0] == 0xFF and (header[1] & 0xE0) == 0xE0):
-                return '.mp3'
-            
-            # OGG: OggS
-            if header[:4] == b'OggS':
-                return '.ogg'
-            
-            # WAV/RIFF: RIFF....WAVE
-            if header[:4] == b'RIFF' and len(header) >= 12 and header[8:12] == b'WAVE':
-                return '.wav'
-            
-            # M4A/AAC: ftyp
-            if len(header) >= 8 and header[4:8] == b'ftyp':
-                return '.m4a'
-            
-            # WMA: ASF header
-            if header[:4] == b'\x30\x26\xb2\x75':
-                return '.wma'
-            
-            logger.debug(f"Unknown audio format, header: {header[:4].hex()}")
-            return '.tmp'
-            
-    except Exception as e:
-        logger.debug(f"Error detecting audio format: {e}")
-        return '.tmp'
-
-def get_all_songs(user_creds=None, provider_type=None, apply_filter=True):
-    """Fetches all songs using admin credentials or explicit creds.
-
-    ``apply_filter`` is forwarded to providers that honor
-    ``config.MUSIC_LIBRARIES`` (currently Navidrome). Migration probes pass
-    ``apply_filter=False`` so the source provider's library filter does not
-    falsely exclude tracks from the target server during dry-run.
-    """
-    provider_type = provider_type or config.MEDIASERVER_TYPE
-    if provider_type == 'jellyfin': return jellyfin_get_all_songs(user_creds=user_creds)
-    if provider_type == 'navidrome': return navidrome_get_all_songs(user_creds=user_creds, apply_filter=apply_filter)
-    if provider_type == 'lyrion': return lyrion_get_all_songs(user_creds=user_creds)
-    if provider_type == 'mpd': return mpd_get_all_songs()
-    if provider_type == 'emby': return emby_get_all_songs(user_creds=user_creds)
-    return []
-
-def list_libraries(user_creds=None, provider_type=None):
-    """List all music libraries/folders a provider exposes.
-
-    Returns {'libraries': [{'id': str, 'name': str}, ...], 'unsupported': bool}.
-    The setup wizard and migration assistant use this to render a checkbox list
-    after a successful test-connection. Uses admin credentials when
-    ``user_creds`` is None, or the supplied creds when probing a target.
-    """
-    provider_type = provider_type or config.MEDIASERVER_TYPE
-    if provider_type == 'jellyfin':  return {'libraries': _jellyfin_list_libraries(user_creds=user_creds), 'unsupported': False}
-    if provider_type == 'navidrome': return {'libraries': _navidrome_list_libraries(user_creds=user_creds), 'unsupported': False}
-    if provider_type == 'lyrion':    return {'libraries': _lyrion_list_libraries(user_creds=user_creds), 'unsupported': False}
-    if provider_type == 'emby':      return {'libraries': _emby_list_libraries(user_creds=user_creds), 'unsupported': False}
-    return {'libraries': [], 'unsupported': True}
-
-def search_albums(query, user_creds=None, provider_type=None):
-    """Searches for albums using admin credentials or explicit creds."""
-    provider_type = provider_type or config.MEDIASERVER_TYPE
-    if provider_type == 'jellyfin': return jellyfin_search_albums(query, user_creds=user_creds)
-    if provider_type == 'navidrome': return navidrome_search_albums(query, user_creds=user_creds)
-    if provider_type == 'lyrion': return lyrion_search_albums(query, user_creds=user_creds)
-    if provider_type == 'mpd': raise NotImplementedError('MPD album search is not supported')
-    if provider_type == 'emby': return emby_search_albums(query, user_creds=user_creds)
-    return []
-
-def test_connection(user_creds=None, provider_type=None):
-    """Tests provider connection using admin credentials or explicit creds."""
-    provider_type = provider_type or config.MEDIASERVER_TYPE
-    if provider_type == 'jellyfin': return jellyfin_test_connection(user_creds=user_creds)
-    if provider_type == 'navidrome': return navidrome_test_connection(user_creds=user_creds)
-    if provider_type == 'lyrion': return lyrion_test_connection(user_creds=user_creds)
-    if provider_type == 'mpd':
-        return {'ok': False, 'error': 'MPD migration probe is not supported', 'sample_count': 0, 'path_format': 'none', 'warnings': []}
-    if provider_type == 'emby': return emby_test_connection(user_creds=user_creds)
-    return {'ok': False, 'error': f"Provider '{provider_type}' not supported", 'sample_count': 0, 'path_format': 'none', 'warnings': []}
-
-def get_playlist_by_name(playlist_name):
-    """Finds a playlist by name using admin credentials."""
-    if not playlist_name: raise ValueError("Playlist name is required.")
-    if config.MEDIASERVER_TYPE == 'jellyfin': return jellyfin_get_playlist_by_name(playlist_name)
-    if config.MEDIASERVER_TYPE == 'navidrome': return navidrome_get_playlist_by_name(playlist_name)
-    if config.MEDIASERVER_TYPE == 'lyrion': return lyrion_get_playlist_by_name(playlist_name)
-    if config.MEDIASERVER_TYPE == 'mpd': return mpd_get_playlist_by_name(playlist_name)
-    if config.MEDIASERVER_TYPE == 'emby': return emby_get_playlist_by_name(playlist_name)
-    return None
-
-def create_playlist(base_name, item_ids):
-    """Creates a playlist using admin credentials."""
-    if not base_name: raise ValueError("Playlist name is required.")
-    if not item_ids: raise ValueError("Track IDs are required.")
-    if config.MEDIASERVER_TYPE == 'jellyfin': jellyfin_create_playlist(base_name, item_ids)
-    elif config.MEDIASERVER_TYPE == 'navidrome': navidrome_create_playlist(base_name, item_ids)
-    elif config.MEDIASERVER_TYPE == 'lyrion': lyrion_create_playlist(base_name, item_ids)
-    elif config.MEDIASERVER_TYPE == 'mpd': mpd_create_playlist(base_name, item_ids)
-    elif config.MEDIASERVER_TYPE == 'emby': emby_create_playlist(base_name, item_ids)
-
-def create_instant_playlist(playlist_name, item_ids, user_creds=None):
-    """Creates an instant playlist. Uses user_creds if provided, otherwise admin."""
-    if not playlist_name: raise ValueError("Playlist name is required.")
-    if not item_ids: raise ValueError("Track IDs are required.")
-
-    if config.MEDIASERVER_TYPE == 'jellyfin':
-        return jellyfin_create_instant_playlist(playlist_name, item_ids, user_creds)
-    if config.MEDIASERVER_TYPE == 'navidrome':
-        return navidrome_create_instant_playlist(playlist_name, item_ids, user_creds)
-    if config.MEDIASERVER_TYPE == 'lyrion':
-        return lyrion_create_instant_playlist(playlist_name, item_ids)
-    if config.MEDIASERVER_TYPE == 'mpd':
-        return mpd_create_instant_playlist(playlist_name, item_ids, user_creds)
-    if config.MEDIASERVER_TYPE == 'emby':
-        return emby_create_instant_playlist(playlist_name, item_ids, user_creds)
-    return None
-
-
-def create_or_replace_playlist(playlist_name, item_ids, user_creds=None):
-    """Cron-only upsert: create the playlist if missing, or replace its contents in place.
-
-    Used by the scheduled sonic_fingerprint task so the same server-side playlist (and ID,
-    where the backend allows) gets reused across runs. Raises NotImplementedError for MPD
-    and any other unsupported backend — the cron handler catches that and falls back to
-    legacy date-suffixed playlist creation.
-    """
-    if not playlist_name:
-        raise ValueError("Playlist name is required.")
-    if not item_ids:
-        raise ValueError("Track IDs are required.")
-
-    if config.MEDIASERVER_TYPE == 'jellyfin':
-        return jellyfin_create_or_replace_playlist(playlist_name, item_ids, user_creds)
-    if config.MEDIASERVER_TYPE == 'navidrome':
-        return navidrome_create_or_replace_playlist(playlist_name, item_ids, user_creds)
-    if config.MEDIASERVER_TYPE == 'emby':
-        return emby_create_or_replace_playlist(playlist_name, item_ids, user_creds)
-    if config.MEDIASERVER_TYPE == 'lyrion':
-        return lyrion_create_or_replace_playlist(playlist_name, item_ids, user_creds)
-    raise NotImplementedError(
-        f"create_or_replace_playlist not supported for MEDIASERVER_TYPE={config.MEDIASERVER_TYPE!r}"
-    )
-
-def get_top_played_songs(limit, user_creds=None):
-    """Fetches top played songs. Uses user_creds if provided, otherwise admin."""
-    if config.MEDIASERVER_TYPE == 'jellyfin':
-        return jellyfin_get_top_played_songs(limit, user_creds)
-    if config.MEDIASERVER_TYPE == 'navidrome':
-        return navidrome_get_top_played_songs(limit, user_creds)
-    if config.MEDIASERVER_TYPE == 'lyrion':
-        return lyrion_get_top_played_songs(limit)
-    if config.MEDIASERVER_TYPE == 'mpd':
-        return mpd_get_top_played_songs(limit, user_creds)
-    if config.MEDIASERVER_TYPE == 'emby':
-        return emby_get_top_played_songs(limit, user_creds)
-    return []
-
-def get_last_played_time(item_id, user_creds=None):
-    """Fetches last played time for a track. Uses user_creds if provided, otherwise admin."""
-    if config.MEDIASERVER_TYPE == 'jellyfin':
-        return jellyfin_get_last_played_time(item_id, user_creds)
-    if config.MEDIASERVER_TYPE == 'navidrome':
-        return navidrome_get_last_played_time(item_id, user_creds)
-    if config.MEDIASERVER_TYPE == 'lyrion':
-        return lyrion_get_last_played_time(item_id)
-    if config.MEDIASERVER_TYPE == 'mpd':
-        return mpd_get_last_played_time(item_id, user_creds)
-    if config.MEDIASERVER_TYPE == 'emby':
-        return emby_get_last_played_time(item_id, user_creds)
-    return None
-
-def get_lyrics(track_id: str, timeout: float = 2.5):
-    """Fetch lyrics embedded in the media server for a given track ID.
-
-    Supported servers: Jellyfin, Emby, Navidrome, Lyrion.
-    MPD does not provide a lyrics API; always returns None.
-    Returns plain text or None.
-    """
-    if config.MEDIASERVER_TYPE == 'jellyfin':
-        return jellyfin_get_lyrics(track_id, timeout=timeout)
-    if config.MEDIASERVER_TYPE == 'emby':
-        return emby_get_lyrics(track_id, timeout=timeout)
-    if config.MEDIASERVER_TYPE == 'navidrome':
-        return navidrome_get_lyrics(track_id, timeout=timeout)
-    if config.MEDIASERVER_TYPE == 'lyrion':
-        return lyrion_get_lyrics(track_id, timeout=timeout)
-    return None
-
