+++ tasks/mediaserver/jellyfin.py
@@ -1,11 +1,11 @@
-# tasks/mediaserver_jellyfin.py
+# tasks/mediaserver/jellyfin.py
 
-from tasks import mediaserver_http as requests
+from . import http as requests
 import logging
 import os
 import config
 
-from tasks.mediaserver_helper import detect_path_format
+from .helper import detect_path_format
 
 logger = logging.getLogger(__name__)
