+++ tasks/mediaserver/navidrome.py
@@ -1,12 +1,12 @@
-# tasks/mediaserver_navidrome.py
+# tasks/mediaserver/navidrome.py
 
-from tasks import mediaserver_http as requests
+from . import http as requests
 import logging
 import os
 import random
 import config
 
-from tasks.mediaserver_helper import detect_path_format
+from .helper import detect_path_format
 
 logger = logging.getLogger(__name__)
